+++ src/lib.rs
@@ -5,5 +5,7 @@ mod generator;
 mod solver;
 mod widget;
 
-pub use game::{hexagon, symmetric_board, Arrows, Dir, GameStatus, HexwebGame, Params, Piece};
+pub use game::{
+    hexagon, random_symmetric_board, Arrows, Dir, GameStatus, HexwebGame, Params, Piece,
+};
 pub use widget::{content_size, fit_cell_size, HexwebWidget};